Move outlines prompts to dedicated file
