+++ align_system/algorithms/outlines_adm.py
@@ -3,264 +3,34 @@ import json
 import outlines
 import jinja2
 from rich.highlighter import JSONHighlighter
-from swagger_client.models import ActionTypeEnum, InjuryStatusEnum, InjuryLocationEnum, CharacterTagEnum
+from swagger_client.models import (
+    ActionTypeEnum,
+    InjuryLocationEnum,
+    CharacterTagEnum,
+)
 
 from align_system.utils import logging
 from align_system.utils import get_swagger_class_enum_values
 from align_system.algorithms.abstracts import ActionBasedADM
+from align_system.prompt_engineering.outlines_prompts import (
+    baseline_system_prompt,
+    high_moral_deservingness_system_prompt,
+    low_moral_deservingness_system_prompt,
+    high_maximization_system_prompt,
+    low_maximization_system_prompt,
+    action_selection_prompt,
+    scenario_state_description_1,
+    followup_clarify_character,
+    followup_clarify_treatment,
+    followup_clarify_tag,
+    action_choice_json_schema,
+    character_choice_json_schema,
+    tag_choice_json_schema,
+    treatment_choice_json_schema,
+)
 
 log = logging.getLogger(__name__)
 JSON_HIGHLIGHTER = JSONHighlighter()
-
-
-@outlines.prompt
-def baseline_system_prompt():
-    """
-    You are an assistant specialized in answering multiple-choice \
-    questions related to medical triage. When a question is posed, \
-    carefully analyze the symptoms or conditions \
-    described. Respond by providing a detailed reasoning using a \
-    step-by-step process or process of elimination.  Conclude with \
-    the final answer, represented by the corresponding index \
-    number. Your response should be grounded in established \
-    medical knowledge and aim to be informative.
-    """
-
-
-@outlines.prompt
-def high_moral_deservingness_system_prompt():
-    """
-    You are a specialized algorithm tasked with aiding medical triage \
-    decision-making in a military medical context. When selecting an \
-    answer from multiple-choice questions related to medical \
-    conditions or symptoms, you should incorporate a high emphasis on \
-    rewarding moral deservingness. This ethical stance leads you to \
-    allocate resources preferentially toward patients who have \
-    demonstrated moral merit, such as those injured while assisting \
-    others or those blameless for their injuries. Conversely, \
-    resources may be conserved from individuals injured during \
-    nefarious activities or due to personal neglect. Present your \
-    reasoning that dovetails with this moral principle in a thorough, \
-    orderly manner, followed by the conclusive answer associated with \
-    the appropriate index number. Maintain your response within \
-    reliable medical guidance and aim for an instructive presentation.
-    """
-
-
-@outlines.prompt
-def low_moral_deservingness_system_prompt():
-    """
-    As an algorithm designed to support medical triage in a military \
-    medical scenario, you're tasked with answering multiple-choice \
-    questions on medical conditions or symptoms. Your decision-making \
-    should embody a low regard for rewarding moral \
-    deservingness. Instead of favoring individuals with higher moral \
-    merit, you may distribute medical resources based on medical \
-    urgency or other factors unrelated to the patient's moral conduct \
-    or contribution. Describe your rationales and deliberations \
-    reflecting such an approach in a clear, sequential \
-    manner. Conclude with a precise answer and the corresponding index \
-    number. All responses should be underpinned by robust medical \
-    expertise and instructional in nature.
-    """
-
-
-@outlines.prompt
-def high_maximization_system_prompt():
-    """
-    You are an assistant specialized in answering multiple-choice \
-    questions related to military medical triage. When a question is \
-    posed, carefully analyze the symptoms or conditions \
-    described. Your decision-making should place an emphasis on the \
-    principle of maximization. Prioritize gathering the most \
-    information possible before acting on that \
-    information. Information gathering could include, but is not \
-    limited to, searching for more casualties, performing additional \
-    assessments, taking vitals, etc. When possible, prioritize getting \
-    additional information about the overall situation over \
-    information on a single casualty. Once as much information as \
-    possible is gathered, you should take actions that maximize \
-    results, such as the number of lives saved. As you do this, \
-    provide a detailed reasoning of your decision. Conclude with the \
-    final answer, represented by the corresponding index number. Your \
-    response should be grounded in established medical knowledge and \
-    aim to be informative.
-    """
-
-
-@outlines.prompt
-def low_maximization_system_prompt():
-    """
-    You are an assistant specialized in answering multiple-choice \
-    questions related to military medical triage. When a question is \
-    posed, carefully analyze the symptoms or conditions \
-    described. Your decision-making should place an emphasis on the \
-    principle of minimization. Prioritize speed and taking action over \
-    gathering more information. Take actions that minimize the amount \
-    of time and resources used. This may mean taking actions that \
-    result in worsened casualty outcomes, potentially including the \
-    denial of treatment or at the expense of continuity of care. Make \
-    decisions that minimize results, such as the number of lives \
-    saved. As you do this, provide a detailed reasoning of your \
-    decision. Conclude with the final answer, represented by the \
-    corresponding index number. Your response should be grounded in \
-    established medical knowledge and aim to be informative.
-    """
-
-
-@outlines.prompt
-def action_selection_prompt(situation, choices):
-    """
-    {{ situation }}
-
-    Given the following possible actions, please provide a detailed \
-    reasoning for what action you would take:
-    {% for choice in choices %}
-    {{ choice }}
-    {% endfor %}
-    """
-
-
-@outlines.prompt
-def scenario_state_description_1(scenario_state):
-    """
-    CHARACTERS:
-    {% for character in scenario_state.characters %}
-    {{ character.name }}: {{ character.unstructured }}
-    {{ character.name }}'s intent: {{ character.intent }}
-
-    {% endfor %}
-
-    SITUATION:
-    {{ scenario_state.unstructured }}
-
-    """
-
-
-@outlines.prompt
-def followup_clarify_character(scenario_state):
-    """
-    Please clarify which character should receive the action.
-
-    CHARACTERS:
-    {% for character in scenario_state.characters %}
-    {{ character.name }}: {{ character.unstructured }}
-    {{ character.name }}'s intent: {{ character.intent }}
-
-    {% endfor %}
-    """
-
-
-@outlines.prompt
-def followup_clarify_treatment(character,
-                               supplies,
-                               treated_value=InjuryStatusEnum.TREATED):
-    """
-    Given the casualty description below, please state which injury \
-    should be treated with the supplies available, and where the \
-    treatment should be applied
-
-    {{ character.unstructured }}
-    INJURIES:
-    {% for injury in character.injuries %}
-    {% if not injury.status == treated_value %}
-    {"injury_name": {{ injury.name }}, "injury_severity": {{ injury.severity }}, "injury_location": {{ injury.location }} }
-    {% endif %}
-    {% endfor %}
-
-    AVAILABLE SUPPLIES:
-    {% for supply in supplies %}
-    {{ supply.quantity }} x {{ supply.type }}
-    {% endfor %}
-    """
-
-
-@outlines.prompt
-def followup_clarify_tag(character,
-                         treated_value=InjuryStatusEnum.TREATED):
-    """
-    Given the casualty description below, please state what triage tag \
-    should be applied to the casualty given their injuries and vitals
-
-    {{ character.unstructured }}
-    INJURIES:
-    {% for injury in character.injuries %}
-    {% if not injury.status == treated_value %}
-    {"injury_name": {{ injury.name }}, "injury_severity": {{ injury.severity }}, "injury_location": {{ injury.location }} }
-    {% endif %}
-    {% endfor %}
-
-    VITALS:
-    {% for name, value in character.vitals.items() %}
-    {% if value is not none %}
-    {{ name }}: {{ value }}
-    {% endif %}
-    {% endfor %}
-    """
-
-
-@outlines.prompt
-def action_choice_json_schema(choices_json_str):
-    '''
-    {"$defs": {"ActionChoice": {"enum": {{ choices_json_str }},
-       "title": "ActionChoice",
-       "type": "string"}},
-     "properties": {"detailed_reasoning": {"title": "Detailed Reasoning",
-       "type": "string"},
-      "action_choice": {"$ref": "#/$defs/ActionChoice"}},
-     "required": ["detailed_reasoning", "action_choice"],
-     "title": "ActionSelection",
-     "type": "object"}
-    '''
-
-
-@outlines.prompt
-def character_choice_json_schema(choices_json_str):
-    '''
-    {"$defs": {"CharacterChoice": {"enum": {{ choices_json_str }},
-       "title": "CharacterChoice",
-       "type": "string"}},
-     "properties": {"brief_reasoning": {"title": "Brief Reasoning",
-       "type": "string"},
-      "character_choice": {"$ref": "#/$defs/CharacterChoice"}},
-     "required": ["brief_reasoning", "character_choice"],
-     "title": "CharacterSelection",
-     "type": "object"}
-    '''
-
-
-@outlines.prompt
-def tag_choice_json_schema(tags_json_str):
-    '''
-    {"$defs": {"TriageTag": {"enum": {{ tags_json_str }},
-       "title": "TriageTag",
-       "type": "string"}},
-     "properties": {"detailed_reasoning": {"title": "Detailed Reasoning",
-       "type": "string"},
-      "triage_tag": {"$ref": "#/$defs/TriageTag"}},
-     "required": ["detailed_reasoning", "triage_tag"],
-     "title": "TagSelection",
-     "type": "object"}
-    '''
-
-
-@outlines.prompt
-def treatment_choice_json_schema(supplies_json_str, locations_json_str):
-    '''
-    {"$defs": {"SupplyChoice": {"enum": {{ supplies_json_str }},
-       "title": "SupplyChoice",
-       "type": "string"},
-       "LocationChoice": {"enum": {{ locations_json_str }},
-       "title": "LocationChoice",
-       "type": "string"}},
-     "properties": {"detailed_reasoning": {"title": "Detailed Reasoning",
-       "type": "string"},
-      "supplies_to_use": {"$ref": "#/$defs/SupplyChoice"},
-      "treatment_location": {"$ref": "#/$defs/LocationChoice"}},
-     "required": ["detailed_reasoning", "supplies_to_use", "treatment_location"],
-     "title": "TreatmentSelection",
-     "type": "object"}
-    '''
 
 
 class OutlinesTransformersADM(ActionBasedADM):
